+++ src/Client.php
@@ -29,7 +29,7 @@ use WebSocket\Exception\{
     BadUriException,
     ClientException,
     ConnectionLevelInterface,
-    Exception,
+    ExceptionInterface,
     HandshakeException,
     MessageLevelInterface,
     ReconnectException
@@ -303,7 +303,7 @@ class Client implements LoggerAwareInterface, Stringable
                 }
                 $this->connection->tick();
                 $this->dispatch('tick', [$this]);
-            } catch (Exception $e) {
+            } catch (ExceptionInterface $e) {
                 $this->disconnect();
                 $this->running = false;
 
@@ -316,7 +316,6 @@ class Client implements LoggerAwareInterface, Stringable
 
                 // Crash it
                 $this->logger->error("[client] {$e->getMessage()}");
-                $this->dispatch('error', [$this, null, $e]);
                 throw $e;
             }
             gc_collect_cycles(); // Collect garbage