+++ docs/Changelog.md
@@ -6,6 +6,11 @@
 
  > PHP version `^8.1`
 
+### `3.2.4`
+
+ * Introducing ExceptionInterface as root to internal exceptions (@sirn-se)
+ * Unresolvable errors must not trigger onError listener (@sirn-se)
+
 ### `3.2.3`
 
  * Allow empty reason phrase in HTTP request (@Sebastix)